+++ strcs/__init__.py
@@ -15,10 +15,7 @@ from .decorator import (
     CreateArgs,
     WrappedCreator,
 )
-from .disassemble.base import Field, Type
-from .disassemble.cache import TypeCache
-from .disassemble.instance_check import InstanceCheck, InstanceCheckMeta
-from .disassemble.type_tree import MRO
+from .disassemble import MRO, Field, InstanceCheck, InstanceCheckMeta, Type, TypeCache
 from .hints import resolve_types
 from .meta import Meta
 from .not_specified import NotSpecified, NotSpecifiedMeta
