+++ strcs/args_extractor.py
@@ -13,7 +13,7 @@ import typing as tp
 
 import cattrs
 
-from .disassemble.base import Type
+from .disassemble import Type
 from .meta import Meta
 
 if tp.TYPE_CHECKING:
