+++ strcs/disassemble/base.py
@@ -3,7 +3,6 @@ import collections.abc
 import functools
 import json
 import operator
-import types
 import typing as tp
 from dataclasses import is_dataclass
 from functools import partial
@@ -16,6 +15,7 @@ from ..hints import resolve_types
 from ..memoized_property import memoized_property
 from ..not_specified import NotSpecifiedMeta
 from ..standard import builtin_types, union_types
+from .extract import IsAnnotated, extract_annotation, extract_optional
 from .fields import Field, fields_from_attrs, fields_from_class, fields_from_dataclasses
 from .instance_check import InstanceCheck, create_checkable
 from .score import Score
@@ -29,44 +29,6 @@ if tp.TYPE_CHECKING:
 
 T = tp.TypeVar("T")
 U = tp.TypeVar("U")
-
-
-class IsAnnotated(tp.Protocol):
-    __args__: tuple
-    __metadata__: tuple
-
-    def copy_with(self, args: tuple) -> type:
-        ...
-
-    @classmethod
-    def has(self, typ: object) -> tp.TypeGuard["IsAnnotated"]:
-        return (
-            tp.get_origin(typ) is tp.Annotated
-            and hasattr(typ, "__args__")
-            and hasattr(typ, "__metadata__")
-        )
-
-
-def extract_optional(typ: T) -> tuple[bool, T]:
-    optional = False
-    if tp.get_origin(typ) in union_types:
-        if type(None) in tp.get_args(typ):
-            optional = True
-
-            remaining = tuple(a for a in tp.get_args(typ) if a not in (types.NoneType,))
-            if len(remaining) == 1:
-                typ = remaining[0]
-            else:
-                typ = functools.reduce(operator.or_, remaining)
-
-    return optional, typ
-
-
-def extract_annotation(typ: T) -> tuple[T, IsAnnotated | None, object | None]:
-    if IsAnnotated.has(typ):
-        return typ.__args__[0], typ, typ.__metadata__[0]
-    else:
-        return typ, None, None
 
 
 @define
