+++ strcs/hints.py
@@ -20,7 +20,7 @@ from attrs import fields as attrs_fields
 from attrs import has as is_attrs
 
 if tp.TYPE_CHECKING:
-    from .disassemble.base import TypeCache
+    from .disassemble import TypeCache
     from .register import CreateRegister
 
 T = tp.TypeVar("T")
@@ -307,7 +307,7 @@ def resolve_types(
                     value = tp.ForwardRef(value, is_argument=False, is_class=True)
 
                 if name in allfields:
-                    from .disassemble.base import Type
+                    from .disassemble import Type
 
                     disassembled = Type.create(value, cache=type_cache, expect=object)
