+++ strcs/hooks.py
@@ -12,9 +12,7 @@ import cattrs
 
 from .annotations import AdjustableCreator, AdjustableMeta
 from .decorator import ConvertDefinition, ConvertFunction, CreateArgs
-from .disassemble.base import Type
-from .disassemble.cache import TypeCache
-from .disassemble.creation import fill, instantiate
+from .disassemble import Type, TypeCache, fill, instantiate
 from .meta import Meta
 from .not_specified import NotSpecified
